SQL migrations added

+++ database/migrations/2023_09_30_191035_create_unlreports_table.php
@@ -1,19 +1,32 @@
 <?php
 
 use Illuminate\Database\Migrations\Migration;
-# use Illuminate\Database\Schema\Blueprint;
-# use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
+
+  public function up()
+  {
+    if(config('xwa.database_engine') == 'bigquery')
+      return $this->up_bigquery();
+    else
+      return $this->up_sql();
+  }
+
+
   /**
    * Run the migrations.
    * Note: BigQuery tables are not instantly available, it takes about 10 seconds to be available.
    * @see https://github.com/prologuetech/laravel-big/blob/master/src/Big.php
    * @return void
    */
-  public function up()
+  protected function up_bigquery()
   {
+    if(config('xwa.database_engine') != 'bigquery')
+      return;
+
     //\BigQuery::createDataset(config('bigquery.xwa_dataset'));
 
     /**
@@ -80,60 +93,24 @@ return new class extends Migration
     ];
 
     \BigQuery::dataset(config('bigquery.xwa_dataset'))->createTable('unlreports', ['schema' => [ 'fields' => $fields ]]);
-    unset($fields);
+    
+  }
 
+  /**
+   * Create accounts table on SQL database.
+   */
+  protected function up_sql()
+  {
+    if(config('xwa.database_engine') != 'sql')
+      return;
 
-    $fields = [
-      [
-        'name' => 'validator',
-        'type' => 'STRING',
-        'mode' => 'REQUIRED',
-        'description' => 'Validator Public Key'
-      ],
-      [
-        'name' => 'account',
-        'type' => 'STRING',
-        'mode' => 'NULLABLE',
-        'description' => 'Account (r-address) of this validator'
-      ],
-      [
-        'name' => 'first_l',
-        'type' => 'INTEGER',
-        'mode' => 'REQUIRED',
-        'description' => 'First appeared ledger index (flag ledger index plus 1)'
-      ],
-      [
-        'name' => 'last_l',
-        'type' => 'INTEGER',
-        'mode' => 'REQUIRED',
-        'description' => 'Last appeared ledger index (flag ledger)'
-      ],
-      [
-        'name' => 'current_successive_fl_count',
-        'type' => 'INTEGER',
-        'mode' => 'REQUIRED',
-        'description' => 'Number of flag ledgers this validator ran without interruptions'
-      ],
-      [
-        'name' => 'max_successive_fl_count',
-        'type' => 'INTEGER',
-        'mode' => 'REQUIRED',
-        'description' => 'Number of flag ledgers this validator ran without interruptions'
-      ],
-      [
-        'name' => 'active_fl_count',
-        'type' => 'INTEGER',
-        'mode' => 'REQUIRED',
-        'description' => 'Number of flag ledgers this validator appeared in'
-      ],
-      /*[
-        'name' => 'total_claimed',
-        'type' => 'INTEGER',
-        'mode' => 'REQUIRED',
-        'description' => 'Claimed drops'
-      ],*/
-    ];
-    \BigQuery::dataset(config('bigquery.xwa_dataset'))->createTable('unlvalidators', ['schema' => [ 'fields' => $fields ]]);
+    Schema::create('unlreports', function (Blueprint $table) {
+      $table->bigIncrements('id');
+      $table->unsignedInteger('first_l')->comment('First applied ledger index (flag ledger index plus 1)');
+      $table->unsignedInteger('last_l')->comment('Last applied ledger index (flag ledger index)');
+      $table->string('vlkey',66)->index()->comment('Validator Public Key'); //eg ED45D1840EE724BE327ABE9146503D5848EFD5F38B6D5FEDE71E80ACCE5E6E738B
+      $table->json('validators')->comment('List of active validators');
+    });
   }
 
   /**
@@ -143,7 +120,31 @@ return new class extends Migration
    */
   public function down()
   {
+    if(config('xwa.database_engine') == 'bigquery')
+      return $this->down_bigquery();
+    else
+      return $this->down_sql();
+  }
+
+  /**
+   * Reverse the migrations.
+   *
+   * @return void
+   */
+  protected function down_bigquery()
+  {
+    if(config('xwa.database_engine') != 'bigquery')
+      return;
+      
     \BigQuery::dataset(config('bigquery.xwa_dataset'))->table('unlreports')->delete();
-    \BigQuery::dataset(config('bigquery.xwa_dataset'))->table('unlvalidators')->delete();
+  }
+
+  protected function down_sql()
+  {
+    if(config('xwa.database_engine') != 'sql')
+      return;
+
+    Schema::dropIfExists('unlreports');
+    
   }
 };
